pipeline/requirements.txt
requests==2.31.0
matplotlib==3.7.1
pandas==2.0.1
boto3==1.26.150
numpy==1.24.3
scikit-learn==1.4.1
PyYAML==6.0
joblib==1.3.1
xgboost==2.0.3
seaborne